page tables: Map 2MiB and 1GiB huge pages
Because we're mapping now the whole pci space into memory we easily run
out of memory because of the size of the page tables. This commit
implements hugetables which can be span 2MiB or 1GiB in size.
Furthermore we introduce easier debuggability by pressing Ctrl+D which
also counts the amount of used heap space.