Mark device tree area on page allocator as reserved
With that we don't need to copy the device tree into a static variable.
We can leave it as it is in memory and just never allocate it.

Yes, this possibly wastes memory because we have to reserve the space
on a page boundary. However, memory is cheap nowadays and it is
preferable to have a simple solution.