Device Tree: Parse without heap allocations
This is necesary if we want to have the amount of total RAM dynamic
and not fixed.

The reason is, that we need to let the page allocator know, how much
memory it has. In order to figure that out, we need to parse the device
tree before the page allocator (and therefore, the heap) is initialized.