mmio: Explicit read and write
Originally I thought that the compiler now reorders mmio accesses because
the output of the system-test was scrambled. However, this was not the
case. This is fixed in bfe8cb151e73e5e036a982bedcd2d6c437b39ddd

Nevertheless, we put a lot of work into that and it is cleaner. So keep
that change.