mmio: Explicit read and write
Originally I thought that the compiler now reorders mmio accesses because
the output of the system-test was scrambled. However, this was not the
case. This is fixed in 178186b30af054b544ba0670aad667564d84c1ff

Nevertheless, we put a lot of work into that and it is cleaner. So keep
that change.