process arguments: Store them on the stack at startup
Store arguments of a process on the newly allocated stack and pass a
pointer to it in the a0 register.

The arguments are stored as C Strings such that we only need to provide
the start of the arguments. The userspace program can figure out the
length and the number of arguments by itself.