Interrupt other CPUs when process has to be killed
It might be that one of the other cpus is currently executing the process
we want to kill. To kill the process immediately we send an interrupt
to all other cpus at the same time. We could figure out to which cpu
we need to send the interrupt, but lets do that at another time.